+++ src/blocks/error-message/__title/error-message__title 2.css
@@ -1,0 +1,19 @@
+.error-message__title {
+    max-width: 510px;
+    color: #1A1B22;
+    font-family: Roboto Slab;
+    font-weight: normal;
+    font-size: 26px;
+    line-height: 30px;
+    margin: auto;
+    margin-top: 0;
+    margin-bottom: 16px;
+}
+
+@media screen and (max-width: 540px) {
+
+    .error-message__title {
+        width: 85%;
+        text-align: center;
+    }
+}
