+++ src/blocks/search-process/__title/search-process__title 2.css
@@ -1,0 +1,10 @@
+.search-process__title {
+    max-width: 197px;
+    font-family: Roboto;
+    font-weight: normal;
+    font-size: 18px;
+    line-height: 24px;
+    margin: auto;
+    margin-top: 0;
+    color: #B6BCBF;
+}
