+++ src/blocks/searcher/__subtitle/searcher__subtitle 2.css
@@ -1,0 +1,42 @@
+.searcher__subtitle {
+    max-width: 580px;
+    font-family: Roboto;
+    font-weight: normal;
+    font-size: 18px;
+    line-height: 24px;
+    color: white;
+    margin-top: 32px;
+    margin-bottom: 0;
+}
+
+@media screen and (max-width: 769px) {
+
+    .searcher__subtitle {
+        max-width: 452px;
+        margin-top: 16px;
+    }
+}
+
+@media screen and (max-width: 483px) {
+
+    .searcher__subtitle {
+        max-width: 400px;
+    }
+}
+
+@media screen and (max-width: 410px) {
+
+    .searcher__subtitle {
+        max-width: 380px;
+    }
+}
+
+@media screen and (max-width: 380px) {
+
+    .searcher__subtitle {
+        font-size: 20px;
+        line-height: 26px;
+        max-width: 288px;
+        margin-top: 56px;
+    }
+}
