+++ src/blocks/slider-card/__date/slider-card__date 2.css
@@ -1,0 +1,9 @@
+.slider-card__date {
+    width: 192px;
+    height: 22px;
+    font-family: Roboto;
+    font-weight: normal;
+    font-size: 14px;
+    line-height: 22px;
+    color: #B6BCBF;
+}
