+++ src/blocks/slider-card/__email/slider-card__email 2.css
@@ -1,0 +1,9 @@
+.slider-card__email {
+    width: 196px;
+    height: 22px;
+    font-family: Roboto;
+    font-weight: normal;
+    font-size: 14px;
+    line-height: 22px;
+    color: #1A1B22;
+}
