+++ src/blocks/slider-card/__photo/slider-card__photo 2.css
@@ -1,0 +1,11 @@
+.slider-card__photo {
+    height: 64px;
+    width: 64px;
+    border-radius: 1000px;
+}
+
+@media screen and (max-width: 1210px) {
+    .slider-card__photo {
+        margin-top: 15px;
+    }
+}
