+++ src/blocks/slider-card/__user-info/slider-card__user-info 2.css
@@ -1,0 +1,6 @@
+.slider-card__user-info {
+    width: 196px;
+    height: 54px;
+    display: flex;
+    flex-direction: column;
+}
