+++ src/blocks/slider-card/__user/slider-card__user 2.css
@@ -1,0 +1,14 @@
+.slider-card__user {
+    height: 64px;
+    width: 276px;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin-top: 16px;
+}
+
+@media screen and (max-width: 1210px) {
+    .slider-card__user {
+        margin-top: 8px;
+    }
+}
