+++ src/blocks/technologies/__cell/technologies__cell 2.css
@@ -1,0 +1,6 @@
+.technologies__cell {
+    width: 90px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
