+++ src/blocks/technologies/__container/technologies__container 2.css
@@ -1,0 +1,11 @@
+.technologies__container{
+    display: flex;
+    justify-content: space-between;
+}
+
+@media screen and (max-width: 980px) {
+
+    .technologies__container{
+        flex-direction: column;
+    }
+}
