+++ src/blocks/technologies/__grid-elem/technologies__grid-elem 2.css
@@ -1,0 +1,5 @@
+.technologies__grid-elem {
+    width: 100%;
+    height: 24px;
+    text-align: center;
+}
