+++ src/blocks/technologies/__grid-icon/technologies__grid-icon 2.css
@@ -1,0 +1,21 @@
+.technologies__grid-icon {
+    width: 88px;
+    height: 97px;
+    margin-top: 23px;
+}
+
+.js-icon {
+    width: 92px;
+    height: 92px
+}
+
+@media screen and (max-width: 570px) {
+
+    .html-icon {
+        margin-bottom: 18px;
+    }
+
+    .css-icon {
+        margin-bottom: 18px;
+    }
+}
