+++ src/blocks/technologies/__grid/technologies__grid 2.css
@@ -1,0 +1,46 @@
+.technologies__grid {
+    max-width: 600px;
+    display: grid;
+    grid-template-columns: repeat(4, 120px);
+    column-gap: 40px;
+    margin-top: 19px;
+}
+
+@media screen and (max-width: 1150px) {
+
+    .technologies__grid {
+        margin-top: 0;
+    }
+}
+
+@media screen and (max-width: 980px) {
+
+    .technologies__grid {
+        margin-top: 40px;
+        column-gap: 64px;
+    }
+}
+
+@media screen and (max-width: 700px) {
+
+    .technologies__grid {
+        column-gap: 40px;
+    }
+}
+
+@media screen and (max-width: 630px) {
+
+    .technologies__grid {
+        column-gap: 20px;
+    }
+}
+
+@media screen and (max-width: 570px) {
+
+    .technologies__grid {
+        grid-template-columns: repeat(2, 120px);
+        row-gap: 41px;
+        margin: 0 auto;
+        margin-top: 40px;
+    }
+}
