+++ src/blocks/technologies/__title/technologies__title 2.css
@@ -1,0 +1,17 @@
+.technologies__title {
+    max-width: 344px;
+    font-family: Roboto Slab;
+    font-weight: normal;
+    font-size: 40px;
+    line-height: 46px;
+    color: #1A1B22;
+    margin: 0;
+}
+
+@media screen and (max-width: 1150px) {
+
+    .technologies__title {
+        font-size: 30px;
+        line-height: 34px;
+    }
+}
