+++ src/blocks/technologies/technologies 2.css
@@ -1,0 +1,68 @@
+@import url('./__cell/technologies__cell.css');
+@import url('./__container/technologies__container.css');
+@import url('./__grid-elem/technologies__grid-elem.css');
+@import url('./__grid-icon/technologies__grid-icon.css');
+@import url('./__grid/technologies__grid.css');
+@import url('./__title/technologies__title.css');
+
+.technologies {
+    background-color: #F5F6F7;
+    padding-left: 104px;
+    padding-right: 230px;
+    padding-top: 75px;
+    padding-bottom: 97px;
+}
+
+@media screen and (max-width: 1301px) {
+
+    .technologies {
+        padding-right: 100px;
+    }
+}
+
+@media screen and (max-width: 1150px) {
+
+    .technologies {
+        padding-right: 50px;
+        padding-left: 50px;
+    }
+}
+
+@media screen and (max-width: 980px) {
+
+    .technologies {
+        padding-top: 40px;
+        padding-left: 80px;
+        padding-bottom: 49px;
+        padding-right: 71px;
+    }
+}
+
+@media screen and (max-width: 780px) {
+
+    .technologies {
+        padding-left: 40px;
+    }
+}
+
+@media screen and (max-width: 570px) {
+
+    .technologies {
+        padding-right: 40px;
+        margin-bottom: 16px;
+    }
+}
+
+@media screen and (max-width: 350px) {
+
+    .technologies {
+        padding-right: 0;
+    }
+}
+
+@media screen and (max-width: 330px) {
+
+    .technologies {
+        padding-left: 20px;
+    }
+}
