+++ src/index-page 2.css
@@ -1,0 +1,11 @@
+@import url('./blocks/about-author/about-author.css');
+@import url('./blocks/search-process/search-process.css');
+@import url('./blocks/cards-container/cards-container.css');
+@import url('./blocks/card/card.css');
+@import url('./blocks/searcher/searcher.css');
+@import url('./blocks/form/form.css');
+@import url('./blocks/header/header.css');
+@import url('./blocks/footer/footer.css');
+@import url('./blocks/no-results/no-results.css');
+@import url('./blocks/page/page.css');
+@import url('./blocks/error-message/error-message.css');
